+++ frameworks/preact/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the library will be documented in this file.
 
+## vX.X.X (Month DD, YYYY)
+
+- Refactor useField and useFieldArray to improve implementation
+
 ## v0.1.0 (July 18, 2025)
 
 - Initial release