+++ frameworks/qwik/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the library will be documented in this file.
 
+## vX.X.X (Month DD, YYYY)
+
+- Refactor useField and useFieldArray to improve implementation
+
 ## v0.1.1 (July 17, 2025)
 
 - Fix unresolvable types of `@formisch/core` and `@formisch/methods` package